vedhvaani-kundali-full/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
pydantic==2.3.0
Jinja2==3.1.2
reportlab==4.0.0
svgwrite==1.4.3
pyswisseph==2.10.3
weasyprint==62.3
